fix crash + collisions

+++ Shelter_Zombies/PlayerAttackComponent.h
@@ -10,7 +10,7 @@ class PlayerAttackComponent : public EntityAttackComponent
 	PlayerAnimationComponent* animation;
 
 public:
-	PlayerAttackComponent(Actor* _owner,const int _damages, const int _range);
+	PlayerAttackComponent(Actor* _owner,const int _damages, const float _range);
 
 public:
 	virtual void Update(const float _deltaTime) override;
